@@ -3,5 +3,5 @@ package common.network;
 /**
  * Created by tacticalsk8er on 9/8/2014.
  */
-public class PacketInOfflineUserUpdate {
+public class PacketInOfflineUserSend {
 }
