@@ -32,8 +32,9 @@ public class NetworkUser {
     private String onlineServer = "";
     private String uuid;
     private boolean disconnected = false;
+    private boolean offline = false;
 
-    private List<Integer> groups = new ArrayList<>();
+    private List<Integer> groups = new ArrayList<>(Arrays.asList(1));
     private boolean banned = false;
     private Date tempBan = null;
     private boolean muted = false;
@@ -55,45 +56,44 @@ public class NetworkUser {
         }
         async_player_map.put(player, this);
         async_uuid_map.put(this.uuid, this);
-        ProxyServer.getInstance().getScheduler().runAsync(VCBungee.getInstance(), new Runnable() {
-            @Override
-            public void run() {
-                DBObject dbObject = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", uuid);
-                if (dbObject != null) {
-                    String groupList = dbObject.get("Group") == null ? "1" : (String) dbObject.get("Group");
-                    groups = parseGroups(groupList);
-                    banned = dbObject.get("Banned") == null ? false : (Boolean) dbObject.get("Banned");
-                    tempBan = (Date) dbObject.get("TempBan");
-                    muted = dbObject.get("Muted") == null ? false : (Boolean) dbObject.get("Muted");
-                    tempMute = (Date) dbObject.get("TempMute");
-                    for (String serverName : VCBungee.servers) {
-                        Object o = dbObject.get(serverName + "-Money");
-                        double value = (o == null ? 0 : (o instanceof Double ? (Double) o : (Integer) o));
-                        double money = dbObject.get(serverName + "-Money") == null ? 0 : value;
-                        moneyList.put(serverName, money);
-                        HashMap<String, String> userdata = dbObject.get(serverName + "-UserData") == null ? new HashMap<String, String>() : parseData((String) dbObject.get(serverName + "-UserData"));
-                        userdataMap.put(serverName, userdata);
-                    }
+        ProxyServer.getInstance().getScheduler().runAsync(VCBungee.getInstance(), () -> {
+            DBObject dbObject = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", uuid);
+            if (dbObject != null) {
+                String groupList = dbObject.get("Group") == null ? "1" : (String) dbObject.get("Group");
+                groups = parseGroups(groupList);
+                banned = dbObject.get("Banned") == null ? false : (Boolean) dbObject.get("Banned");
+                tempBan = (Date) dbObject.get("TempBan");
+                muted = dbObject.get("Muted") == null ? false : (Boolean) dbObject.get("Muted");
+                tempMute = (Date) dbObject.get("TempMute");
+                for (String serverName : VCBungee.servers) {
+                    Object o = dbObject.get(serverName + "-Money");
+                    double value = (o == null ? 0 : (o instanceof Double ? (Double) o : (Integer) o));
+                    double money = dbObject.get(serverName + "-Money") == null ? 0 : value;
+                    moneyList.put(serverName, money);
+                    HashMap<String, String> userdata = dbObject.get(serverName + "-UserData") == null ? new HashMap<>() : parseData((String) dbObject.get(serverName + "-UserData"));
+                    userdataMap.put(serverName, userdata);
+                }
 
-                    tokens = dbObject.get("Tokens") == null ? 0 : (Integer) dbObject.get("Tokens");
-                    globalUserdata = dbObject.get("Global-UserData") == null ? new HashMap<String, String>() : parseData((String) dbObject.get("Global-UserData"));
-                } else {
-                    for (String serverName : VCBungee.servers) {
-                        moneyList.put(serverName, 0.0);
-                        userdataMap.put(serverName, new HashMap<String, String>());
-                    }
+                tokens = dbObject.get("Tokens") == null ? 0 : (Integer) dbObject.get("Tokens");
+                globalUserdata = dbObject.get("Global-UserData") == null ? new HashMap<>() : parseData((String) dbObject.get("Global-UserData"));
+            } else {
+                for (String serverName : VCBungee.servers) {
+                    moneyList.put(serverName, 0.0);
+                    userdataMap.put(serverName, new HashMap<>());
                 }
-                if(WhitelistCommand.isWhitelisted()) {
-                    if(!GroupUtil.hasPermission(groups, GroupUtil.Group.ADMIN) && WhitelistCommand.isHardMode() && !WhitelistCommand.getWhitelist().contains(uuid)) {
-                            player.disconnect(new TextComponent("You are not whitelisted!"));
-                            return;
-                    } else if(!GroupUtil.hasPermission(groups, GroupUtil.Group.HELPER) && !WhitelistCommand.getWhitelist().contains(uuid)) {
-                        player.disconnect(new TextComponent("You are not whitelisted!"));
-                        return;
-                    }
-                }
-                UserReadyThread.addReadyUser(NetworkUser.this);
             }
+            if (WhitelistCommand.isWhitelisted()) {
+                if (!GroupUtil.hasPermission(groups, GroupUtil.Group.ADMIN) && WhitelistCommand.isHardMode() && !WhitelistCommand.getWhitelist().contains(uuid)) {
+                    player.disconnect(new TextComponent("You are not whitelisted!"));
+                    NetworkUser.remove(player);
+                    return;
+                } else if (!GroupUtil.hasPermission(groups, GroupUtil.Group.HELPER) && !WhitelistCommand.getWhitelist().contains(uuid)) {
+                    player.disconnect(new TextComponent("You are not whitelisted!"));
+                    NetworkUser.remove(player);
+                    return;
+                }
+            }
+            UserReadyThread.addReadyUser(NetworkUser.this);
         });
     }
 
@@ -178,13 +178,13 @@ public class NetworkUser {
     }
 
     public void setMoney(String serverName, double money) {
-        if(moneyList.containsKey(serverName))
+        if (moneyList.containsKey(serverName))
             moneyList.remove(serverName);
         moneyList.put(serverName, money);
     }
 
     public void setUserdata(String serverName, HashMap<String, String> userdata) {
-        if(userdataMap.containsKey(serverName))
+        if (userdataMap.containsKey(serverName))
             userdataMap.remove(serverName);
         userdataMap.put(serverName, userdata);
     }
@@ -198,57 +198,52 @@ public class NetworkUser {
     }
 
     public void save() {
-        ProxyServer.getInstance().getScheduler().runAsync(VCBungee.getInstance(), new Runnable() {
-            @Override
-            public void run() {
-                DBObject dbObject = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", getUuid()) == null ? new BasicDBObject() : VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", getUuid());
-                dbObject.put("UUID", getUuid());
-                dbObject.put("Group", groupsToString(getGroups()));
-                dbObject.put("Banned", isBanned());
-                dbObject.put("TempBan", getTempBan());
-                dbObject.put("Muted", isMuted());
-                dbObject.put("TempMute", getTempMute());
-                for (String serverName : VCBungee.servers) {
-                    dbObject.put(serverName + "-Money", getMoney(serverName));
-                    dbObject.put(serverName + "-UserData", dataToString(getUserdata(serverName)));
-                }
-                dbObject.put("Tokens", getTokens());
-                dbObject.put("Global-UserData", dataToString(getGlobalUserdata()));
-                DBObject dbObject1 = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", getUuid());
-                if (dbObject1 == null)
-                    VCBungee.getInstance().getMongoDB().insert("VaultCraft", "Users", dbObject);
-                else
-                    VCBungee.getInstance().getMongoDB().update("VaultCraft", "Users", dbObject1, dbObject);
+        ProxyServer.getInstance().getScheduler().runAsync(VCBungee.getInstance(), () -> {
+            DBObject dbObject = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", getUuid()) == null ? new BasicDBObject() : VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", getUuid());
+            dbObject.put("UUID", getUuid());
+            dbObject.put("Group", groupsToString(getGroups()));
+            dbObject.put("Banned", isBanned());
+            dbObject.put("TempBan", getTempBan());
+            dbObject.put("Muted", isMuted());
+            dbObject.put("TempMute", getTempMute());
+            for (String serverName : VCBungee.servers) {
+                dbObject.put(serverName + "-Money", getMoney(serverName));
+                dbObject.put(serverName + "-UserData", dataToString(getUserdata(serverName)));
             }
+            dbObject.put("Tokens", getTokens());
+            dbObject.put("Global-UserData", dataToString(getGlobalUserdata()));
+            DBObject dbObject1 = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", getUuid());
+            if (dbObject1 == null)
+                VCBungee.getInstance().getMongoDB().insert("VaultCraft", "Users", dbObject);
+            else
+                VCBungee.getInstance().getMongoDB().update("VaultCraft", "Users", dbObject1, dbObject);
         });
     }
 
     public static void remove(final ProxiedPlayer player) {
         final NetworkUser user = async_player_map.get(player);
-        ProxyServer.getInstance().getScheduler().runAsync(VCBungee.getInstance(), new Runnable() {
-            @Override
-            public void run() {
-                DBObject dbObject = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", user.getUuid()) == null ? new BasicDBObject() : VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", user.getUuid());
-                dbObject.put("UUID", user.getUuid());
-                dbObject.put("Group", groupsToString(user.getGroups()));
-                dbObject.put("Banned", user.isBanned());
-                dbObject.put("TempBan", user.getTempBan());
-                dbObject.put("Muted", user.isMuted());
-                dbObject.put("TempMute", user.getTempMute());
-                for (String serverName : VCBungee.servers) {
-                    dbObject.put(serverName + "-Money", user.getMoney(serverName));
-                    dbObject.put(serverName + "-UserData", dataToString(user.getUserdata(serverName)));
-                }
-                dbObject.put("Tokens", user.getTokens());
-                dbObject.put("Global-UserData", dataToString(user.getGlobalUserdata()));
-                DBObject dbObject1 = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", user.getUuid());
-                if (dbObject1 == null)
-                    VCBungee.getInstance().getMongoDB().insert("VaultCraft", "Users", dbObject);
-                else
-                    VCBungee.getInstance().getMongoDB().update("VaultCraft", "Users", dbObject1, dbObject);
+        ProxyServer.getInstance().getScheduler().runAsync(VCBungee.getInstance(), () -> {
+            DBObject dbObject = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", user.getUuid()) == null ? new BasicDBObject() : VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", user.getUuid());
+            dbObject.put("UUID", user.getUuid());
+            dbObject.put("Group", groupsToString(user.getGroups()));
+            dbObject.put("Banned", user.isBanned());
+            dbObject.put("TempBan", user.getTempBan());
+            dbObject.put("Muted", user.isMuted());
+            dbObject.put("TempMute", user.getTempMute());
+            for (String serverName : VCBungee.servers) {
+                dbObject.put(serverName + "-Money", user.getMoney(serverName));
+                dbObject.put(serverName + "-UserData", dataToString(user.getUserdata(serverName)));
             }
+            dbObject.put("Tokens", user.getTokens());
+            dbObject.put("Global-UserData", dataToString(user.getGlobalUserdata()));
+            DBObject dbObject1 = VCBungee.getInstance().getMongoDB().query("VaultCraft", "Users", "UUID", user.getUuid());
+            if (dbObject1 == null)
+                VCBungee.getInstance().getMongoDB().insert("VaultCraft", "Users", dbObject);
+            else
+                VCBungee.getInstance().getMongoDB().update("VaultCraft", "Users", dbObject1, dbObject);
         });
         async_player_map.remove(player);
+        async_uuid_map.remove(user.getUuid());
     }
 
     public static void disable() {
@@ -290,7 +285,7 @@ public class NetworkUser {
         HashMap<String, String> userdata = new HashMap<>();
         if (!(data.contains("▲")))
             return userdata;
-        if(data.contains("▼")) {
+        if (data.contains("▼")) {
             String[] parts = data.split("▼");
             for (String s : parts) {
                 String[] entry = s.split("▲");
@@ -306,10 +301,10 @@ public class NetworkUser {
     private static List<Integer> parseGroups(String s) {
         List<Integer> groups = new ArrayList<>();
         String[] parts = s.split(",");
-        for(String part : parts) {
+        for (String part : parts) {
             try {
                 groups.add(Integer.parseInt(part));
-            } catch(NumberFormatException ignored) {
+            } catch (NumberFormatException ignored) {
             }
         }
         return groups;
@@ -317,8 +312,8 @@ public class NetworkUser {
 
     private static String groupsToString(List<Integer> groups) {
         StringBuilder sb = new StringBuilder();
-        for(int i = 0; i < groups.size(); i++) {
-            if(i + 1 == groups.size())
+        for (int i = 0; i < groups.size(); i++) {
+            if (i + 1 == groups.size())
                 sb.append(groups.get(i));
             else
                 sb.append(groups.get(i)).append(",");
